Initial Commit
SiteLocator has been modified from the original application.

Added location timer to indicate that a location has been selected, and the application is waiting for a response from the location service.
Click events on the map will be ignored while the application is waiting for location information to be returned from the location service.
Added a named span to the 12 digit watershed information to allow the data to be scraped.